dont render background music on ios devices
